Move types module into its own directory
