+++ crates/nvue-client/src/types/mod.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, serde::Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub struct MacTableEntry {
+    pub mac: String,
+    pub interface: String,
+    pub entry_type: String,
+    pub vlan: Option<u16>,
+}
